requirements.txt
aiogram==3.*
python-dotenv>=1.0
aiohttp>=3.9
aiosqlite>=0.19
